+++ spec/models/request_spec.rb
@@ -3,6 +3,13 @@
 require 'rails_helper'
 
 describe Doorkeeper::OpenidConnect::Request do
+  let(:expected_access_grant_class_name) do
+    if Gem.loaded_specs['doorkeeper'].version >= Gem::Version.create('5.5.0')
+      Doorkeeper.config.access_grant_class.to_s
+    else
+      'Doorkeeper::AccessGrant'
+    end
+  end
   describe 'validations' do
     it 'requires an access grant' do
       subject.access_grant_id = nil
@@ -24,7 +31,7 @@ describe Doorkeeper::OpenidConnect::Request do
       association = subject.class.reflect_on_association :access_grant
 
       expect(association.options).to eq({
-        class_name: 'Doorkeeper::AccessGrant',
+        class_name: expected_access_grant_class_name,
         inverse_of: :openid_request,
       })
     end